requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
flask>=2.3.0
flask-cors>=4.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0

